pyproject.toml
[project]
name = "voice-agent"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "livekit-agents[openai]~=1.4",
    "livekit-plugins-deepgram~=1.0",
    "livekit-plugins-noise-cancellation~=0.2",
    "google-api-python-client>=2.176.0",
    "google-auth>=2.40.0",
    "python-dotenv>=1.2.1",
]
